requirements/production.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   openai
asgiref==3.10.0
    # via django
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via cryptography
charset-normalizer==3.4.3
    # via requests
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   loguru
    #   pytest
    #   tqdm
coverage==7.10.7
    # via
    #   django-coverage-plugin
    #   pytest-cov
cryptography==46.0.2
    # via pyjwt
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
distro==1.9.0
    # via openai
django==5.2.7
    # via
    #   cosmos
    #   django-coverage-plugin
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   django-model-utils
    #   django-silk
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   djoser
    #   drf-spectacular
    #   social-auth-app-django
django-coverage-plugin==3.2.0
django-debug-toolbar==6.0.0
django-environ==0.12.0
    # via cosmos
django-extensions==4.1
    # via cosmos
django-filter==25.2
    # via cosmos
django-model-utils==5.0.0
    # via cosmos
django-silk==5.4.3
djangorestframework==3.16.1
    # via
    #   cosmos
    #   djangorestframework-simplejwt
    #   djoser
    #   drf-spectacular
djangorestframework-simplejwt==5.5.1
    # via
    #   cosmos
    #   djoser
djoser==2.3.3
    # via cosmos
drf-spectacular==0.28.0
    # via cosmos
execnet==2.1.1
    # via pytest-xdist
factory-boy==3.3.3
    # via cosmos
faker==37.11.0
    # via factory-boy
gprof2dot==2025.4.14
    # via django-silk
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via openai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
inflection==0.5.1
    # via drf-spectacular
iniconfig==2.1.0
    # via pytest
jiter==0.11.0
    # via openai
jsonschema==4.25.1
    # via drf-spectacular
jsonschema-specifications==2025.9.1
    # via jsonschema
loguru==0.7.3
    # via cosmos
mysqlclient==2.2.7
    # via cosmos
oauthlib==3.3.1
    # via
    #   requests-oauthlib
    #   social-auth-core
openai==2.3.0
    # via cosmos
packaging==25.0
    # via pytest
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
pycparser==2.23 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
pydantic==2.12.0
    # via openai
pydantic-core==2.41.1
    # via pydantic
pygments==2.19.2
    # via pytest
pyjwt==2.10.1
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
pytest==8.4.2
    # via
    #   pytest-cov
    #   pytest-django
    #   pytest-env
    #   pytest-sugar
    #   pytest-xdist
pytest-cov==7.0.0
pytest-django==4.11.1
pytest-env==1.2.0
pytest-sugar==1.1.1
pytest-xdist==3.8.0
python3-openid==3.2.0
    # via social-auth-core
pyyaml==6.0.3
    # via
    #   drf-spectacular
    #   responses
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   requests-oauthlib
    #   responses
    #   social-auth-core
requests-oauthlib==2.0.0
    # via social-auth-core
responses==0.25.8
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
ruff==0.14.0
six==1.17.0
    # via transitions
sniffio==1.3.1
    # via
    #   anyio
    #   openai
social-auth-app-django==5.6.0
    # via djoser
social-auth-core==4.8.1
    # via social-auth-app-django
sqlparse==0.5.3
    # via
    #   django
    #   django-debug-toolbar
    #   django-silk
stackprinter==0.2.12
    # via cosmos
termcolor==3.1.0
    # via pytest-sugar
tqdm==4.67.1
    # via openai
transitions==0.9.3
    # via cosmos
typing-extensions==4.15.0
    # via
    #   anyio
    #   openai
    #   pydantic
    #   pydantic-core
    #   referencing
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via
    #   django
    #   faker
uritemplate==4.2.0
    # via drf-spectacular
urllib3==2.5.0
    # via
    #   requests
    #   responses
win32-setctime==1.2.0 ; sys_platform == 'win32'
    # via loguru
